refactor: Calendar 관련 타입 추가, 컴포넌트 분리
